Homework

+++ src/main/java/io/github/bael/spring/data/service/BookTransactionService.java
@@ -4,12 +4,14 @@ import io.github.bael.spring.data.entity.Book;
 import io.github.bael.spring.data.entity.Customer;
 import io.github.bael.spring.data.entity.PurchasedBook;
 
+import java.math.BigDecimal;
+
 
 public interface BookTransactionService {
     void save(Book book);
     void save(Customer customer);
-    void doTransaction(Customer customer, Book book, double cost);
-    Double sumOfBookSales(Book book);
-    Double sumOfCostPurchasedBooksByCustomer(Customer customer);
+    void doTransaction(Customer customer, Book book, BigDecimal cost);
+    BigDecimal sumOfBookSales(Book book);
+    BigDecimal sumOfCostPurchasedBooksByCustomer(Customer customer);
 
 }
